requirements.txt
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

transformers>=4.40.0
accelerate>=0.28.0

numpy>=1.26
pandas>=2.2.0

tqdm>=4.66.0
fastapi
scikit-learn>=1.3.0
rembg
Pillow>=10.0.0
open-clip-torch
gymnasium>=0.29.1
stable-baselines3>=2.3.0
openai
mcp
rich>=13.7.0
onnxruntime
deep_translator